requirements.txt
# Core Dependencies
pandas>=1.5.0
numpy>=1.21.0
yfinance>=0.2.0
requests>=2.28.0
python-dotenv>=0.19.0

# Machine Learning
scikit-learn>=1.1.0
joblib>=1.2.0

# Google Sheets Integration
google-auth>=2.15.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0
gspread>=5.7.0

# Telegram Integration
python-telegram-bot>=20.0

# Data Visualization (Optional)
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.12.0

# Development Dependencies
black>=22.0.0
flake8>=5.0.0
pylint>=2.15.0
mypy>=0.991
isort>=5.10.0
pre-commit>=2.20.0

# Testing Dependencies
pytest>=7.2.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.20.0

# Documentation Dependencies
sphinx>=5.3.0
sphinx-rtd-theme>=1.1.0
mkdocs>=1.4.0
mkdocs-material>=8.5.0

# Security and Code Quality
bandit>=1.7.0
safety>=2.3.0

# Performance and Monitoring
memory-profiler>=0.60.0
line-profiler>=4.0.0

# Jupyter Notebooks (Optional)
jupyter>=1.0.0
ipykernel>=6.17.0